🚛 Move components in their own directory 🚛
