+++ src/components/layout/Layout.component.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import Helmet from 'react-helmet'
 import { StaticQuery, graphql } from 'gatsby'
 
-import Header from './header'
+import Header from '../Header'
 import './layout.css'
 import '@ibm/plex/css/ibm-plex.min.css'
